Move all YouTube files into the same namespace
