+++ lib/Destiny/Youtube/YouTubeBroadcasterAuthHandler.php
@@ -1,5 +1,5 @@
 <?php
-namespace Destiny\Google;
+namespace Destiny\Youtube;
 
 use Destiny\Common\Authentication\AuthProvider;
 use Destiny\Common\Authentication\OAuthResponse;
@@ -8,6 +8,7 @@ use Destiny\Common\Exception;
 use Destiny\Common\Session\Session;
 use Destiny\Common\Utils\FilterParams;
 use Destiny\Common\Utils\Http;
+use Destiny\Google\GoogleAuthHandler;
 
 class YouTubeBroadcasterAuthHandler extends GoogleAuthHandler {
     private $apiBase = 'https://www.googleapis.com/youtube/v3';
